Fixed find_track bug
Fixed a bug where the track search would fail if a track was searched in
the wrong album when a user input an album name that was a substring of
multiple albums.

This fix allows the search to continue in other albums if the search
fails in the current album.